HomeController logic refactorization
